Reorganize the codebase into concern-based components and move views to templates
