+++ includes/Admin/Menus.php
@@ -2,9 +2,9 @@
 
 namespace WooCommerceSerialNumbers\Admin;
 
+use WooCommerceSerialNumbers\Activations;
 use WooCommerceSerialNumbers\B8\Component;
-
-use WooCommerceSerialNumbers\Models\Key;
+use WooCommerceSerialNumbers\Keys;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -15,10 +15,12 @@ defined( 'ABSPATH' ) || exit;
  * @package WooCommerceSerialNumbers\Admin
  */
 class Menus extends Component {
+
 	/**
 	 * Register hooks.
 	 *
 	 * @since 1.0.0
+	 * @return void
 	 */
 	public function register(): void {
 		// Register the menus.
@@ -29,54 +31,10 @@ class Menus extends Component {
 		add_action( 'admin_menu', array( $this, 'settings_menu' ), 100 );
 		add_action( 'admin_menu', array( $this, 'promo_menu' ), PHP_INT_MAX );
 
-		// Add tabs content.
-		add_filter( 'wc_serial_numbers_tools_tabs', array( __CLASS__, 'add_tools_status_tab' ), PHP_INT_MAX );
+		// The pro plugin removes these tabs by the static callable names, so they must stay static.
 		add_action( 'wc_serial_numbers_tools_tab_import', array( __CLASS__, 'import_tab' ) );
 		add_action( 'wc_serial_numbers_tools_tab_export', array( __CLASS__, 'export_tab' ) );
 		add_action( 'wc_serial_numbers_tools_tab_generators', array( __CLASS__, 'generators_tab' ) );
-		add_action( 'wc_serial_numbers_tools_tab_status', array( __CLASS__, 'status_tab' ) );
-		add_action( 'wc_serial_numbers_tools_tab_api', array( __CLASS__, 'api_validation_section' ) );
-		add_action( 'wc_serial_numbers_tools_tab_api', array( __CLASS__, 'api_activation_deactivation_section' ) );
-		add_action( 'wc_serial_numbers_reports_tab_stock', array( __CLASS__, 'reports_stock_tab' ) );
-	}
-
-	/**
-	 * Looks at the current screen and loads the correct list table handler.
-	 *
-	 * @since 1.4.6
-	 */
-	public function setup_screen() {
-		wp_verify_nonce( '_nonce' );
-
-		if ( isset( $_GET['edit'] ) || isset( $_GET['delete'] ) || isset( $_GET['add'] ) || isset( $_GET['generate'] ) ) {
-			return;
-		}
-
-		$screen_id        = false;
-		$plugin_screen_id = sanitize_title( __( 'Serial Numbers', 'wc-serial-numbers' ) );
-		if ( function_exists( 'get_current_screen' ) ) {
-			$screen    = get_current_screen();
-			$screen_id = isset( $screen, $screen->id ) ? $screen->id : '';
-		}
-
-		// Ensure the table handler is only loaded once. Prevents multiple loads if a plugin calls check_ajax_referer many times.
-		remove_action( 'current_screen', array( $this, 'setup_screen' ) );
-		remove_action( 'check_ajax_referer', array( $this, 'setup_screen' ) );
-	}
-
-	/**
-	 * Validate screen options on update.
-	 *
-	 * @param bool|int $status Screen option value. Default false to skip.
-	 * @param string   $option The option name.
-	 * @param int      $value The number of rows to use.
-	 */
-	public function save_screen_options( $status, $option, $value ) {
-		if ( in_array( $option, array( 'wsn_keys_per_page', 'wsn_generators_per_page', 'wsn_activations_per_page' ), true ) ) {
-			return $value;
-		}
-
-		return $status;
 	}
 
 	/**
@@ -102,7 +60,7 @@ class Menus extends Component {
 			__( 'Serial Keys', 'wc-serial-numbers' ),
 			'manage_woocommerce', // phpcs:ignore WordPress.WP.Capabilities.Unknown
 			'wc-serial-numbers',
-			array( $this, 'output_main_page' )
+			$this->app->callback( array( Keys\Admin::class, 'output_page' ) )
 		);
 	}
 
@@ -122,7 +80,7 @@ class Menus extends Component {
 			__( 'Activations', 'wc-serial-numbers' ),
 			'manage_woocommerce', // phpcs:ignore WordPress.WP.Capabilities.Unknown
 			'wc-serial-numbers-activations',
-			array( $this, 'output_activations_page' )
+			$this->app->callback( array( Activations\Admin::class, 'output_page' ) )
 		);
 	}
 
@@ -139,7 +97,7 @@ class Menus extends Component {
 			__( 'Tools', 'wc-serial-numbers' ),
 			'manage_woocommerce', // phpcs:ignore WordPress.WP.Capabilities.Unknown
 			'wc-serial-numbers-tools',
-			array( $this, 'output_tools_page' )
+			$this->app->callback( array( Tools::class, 'output_page' ) )
 		);
 	}
 
@@ -156,7 +114,7 @@ class Menus extends Component {
 			__( 'Reports', 'wc-serial-numbers' ),
 			'manage_woocommerce', // phpcs:ignore WordPress.WP.Capabilities.Unknown
 			'wc-serial-numbers-reports',
-			array( $this, 'output_reports_page' )
+			$this->app->callback( array( Reports::class, 'output_page' ) )
 		);
 	}
 
@@ -197,107 +155,6 @@ class Menus extends Component {
 	}
 
 	/**
-	 * Output keys page.
-	 *
-	 * @since 1.0.0
-	 * @return void
-	 */
-	public function output_main_page() {
-		wp_verify_nonce( '_nonce' );
-		$add  = isset( $_GET['add'] ) ? true : false;
-		$edit = isset( $_GET['edit'] ) ? absint( $_GET['edit'] ) : 0;
-		if ( $edit ) {
-			$key = new Key( $edit );
-			if ( ! $key->exists() ) {
-				wp_safe_redirect( remove_query_arg( 'edit' ) );
-				exit();
-			}
-		}
-
-		if ( $add ) {
-			$key = new Key();
-			include __DIR__ . '/views/html-edit-key.php';
-		} elseif ( $edit ) {
-			include __DIR__ . '/views/html-edit-key.php';
-		} else {
-			include __DIR__ . '/views/html-list-keys.php';
-		}
-	}
-
-	/**
-	 * Output activations page.
-	 *
-	 * @since 1.0.0
-	 * @return void
-	 */
-	public function output_activations_page() {
-		Admin::view( 'html-list-activations.php' );
-	}
-
-
-	/**
-	 * Output tools page.
-	 *
-	 * @since 1.0.0
-	 * @return void
-	 */
-	public function output_tools_page() {
-		wp_verify_nonce( '_nonce' );
-		$tabs = array(
-			'generators' => __( 'Generators', 'wc-serial-numbers' ),
-			'api'        => __( 'API Toolkit', 'wc-serial-numbers' ),
-			'import'     => __( 'Import', 'wc-serial-numbers' ),
-			'export'     => __( 'Export', 'wc-serial-numbers' ),
-		);
-
-		// If software support is disabled, remove the activations tab.
-		if ( ! wcsn_is_software_support_enabled() ) {
-			unset( $tabs['api'] );
-		}
-
-		$tabs        = apply_filters( 'wc_serial_numbers_tools_tabs', $tabs );
-		$tab_ids     = array_keys( $tabs );
-		$current_tab = isset( $_GET['tab'] ) ? sanitize_key( wp_unslash( $_GET['tab'] ) ) : reset( $tab_ids );
-		$page        = isset( $_GET['page'] ) ? sanitize_key( wp_unslash( $_GET['page'] ) ) : '';
-
-		Admin::view(
-			'html-tools.php',
-			array(
-				'tabs'        => $tabs,
-				'current_tab' => $current_tab,
-				'page'        => $page,
-			)
-		);
-	}
-
-	/**
-	 * Output reports page.
-	 *
-	 * @since 1.0.0
-	 * @return void
-	 */
-	public function output_reports_page() {
-		wp_verify_nonce( '_nonce' );
-		$tabs = array(
-			'stock' => __( 'Stock', 'wc-serial-numbers' ),
-		);
-
-		$tabs        = apply_filters( 'wc_serial_numbers_reports_tabs', $tabs );
-		$tab_ids     = array_keys( $tabs );
-		$current_tab = isset( $_GET['tab'] ) ? sanitize_key( wp_unslash( $_GET['tab'] ) ) : reset( $tab_ids );
-		$page        = isset( $_GET['page'] ) ? sanitize_key( wp_unslash( $_GET['page'] ) ) : '';
-
-		Admin::view(
-			'html-reports.php',
-			array(
-				'tabs'        => $tabs,
-				'current_tab' => $current_tab,
-				'page'        => $page,
-			)
-		);
-	}
-
-	/**
 	 * Redirect to pro page.
 	 *
 	 * @since 1.0.0
@@ -312,41 +169,13 @@ class Menus extends Component {
 	}
 
 	/**
-	 * Add status tab.
-	 *
-	 * @param array $tabs Tabs.
-	 *
-	 * @return array
-	 */
-	public static function add_tools_status_tab( $tabs ) {
-		$tabs['status'] = __( 'Status', 'wc-serial-numbers' );
-
-		return $tabs;
-	}
-
-	/**
 	 * Import tab content.
 	 *
 	 * @since 1.0.0
 	 * @return void
 	 */
 	public static function import_tab() {
-		?>
-		<div class="wcsn-feature-promo-banner">
-			<div class="wcsn-feature-promo-banner__content">
-				<h3><?php esc_html_e( 'Available in Pro Version', 'wc-serial-numbers' ); ?></h3>
-				<a href="https://pluginever.com/plugins/woocommerce-serial-numbers-pro/?utm_source=import-tab&utm_medium=link&utm_campaign=upgrade&utm_id=wc-serial-numbers" target="_blank" class="button-primary"><?php esc_html_e( 'Upgrade to Pro Now', 'wc-serial-numbers' ); ?></a>
-			</div>
-			<img src="<?php echo esc_url( WCSN()->assets_url( 'images/csv-import.png' ) ); ?>" alt="<?php esc_attr_e( 'Import Serial Numbers', 'wc-serial-numbers' ); ?>"/>
-		</div>
-		<div class="wcsn-feature-promo-banner">
-			<div class="wcsn-feature-promo-banner__content">
-				<h3><?php esc_html_e( 'Available in Pro Version', 'wc-serial-numbers' ); ?></h3>
-				<a href="https://pluginever.com/plugins/woocommerce-serial-numbers-pro/?utm_source=import-tab&utm_medium=link&utm_campaign=upgrade&utm_id=wc-serial-numbers" target="_blank" class="button-primary"><?php esc_html_e( 'Upgrade to Pro Now', 'wc-serial-numbers' ); ?></a>
-			</div>
-			<img src="<?php echo esc_url( WCSN()->assets_url( 'images/txt-import.png' ) ); ?>" alt="<?php esc_attr_e( 'Import Serial Numbers', 'wc-serial-numbers' ); ?>"/>
-		</div>
-		<?php
+		WCSN()->make( Tools::class )->import_tab();
 	}
 
 	/**
@@ -356,163 +185,16 @@ class Menus extends Component {
 	 * @return void
 	 */
 	public static function export_tab() {
-		?>
-		<div class="wcsn-feature-promo-banner">
-			<div class="wcsn-feature-promo-banner__content">
-				<h3><?php esc_html_e( 'Available in Pro Version', 'wc-serial-numbers' ); ?></h3>
-				<a href="https://pluginever.com/plugins/woocommerce-serial-numbers-pro/?utm_source=export-tab&utm_medium=link&utm_campaign=upgrade&utm_id=wc-serial-numbers" target="_blank" class="button-primary"><?php esc_html_e( 'Upgrade to Pro Now', 'wc-serial-numbers' ); ?></a>
-			</div>
-			<img src="<?php echo esc_url( WCSN()->assets_url( 'images/csv-export.png' ) ); ?>" alt="<?php esc_attr_e( 'Export Serial Numbers', 'wc-serial-numbers' ); ?>"/>
-		</div>
-		<?php
+		WCSN()->make( Tools::class )->export_tab();
 	}
 
 	/**
-	 * Getnerators tab content.
+	 * Generators tab content.
 	 *
 	 * @since 1.4.6
 	 * @return void
 	 */
 	public static function generators_tab() {
-		?>
-		<div class="wcsn-feature-promo-banner">
-			<div class="wcsn-feature-promo-banner__content">
-				<h3><?php esc_html_e( 'Available in Pro Version', 'wc-serial-numbers' ); ?></h3>
-				<a href="https://pluginever.com/plugins/woocommerce-serial-numbers-pro/?utm_source=generators-tab&utm_medium=link&utm_campaign=upgrade&utm_id=wc-serial-numbers" target="_blank" class="button-primary"><?php esc_html_e( 'Upgrade to Pro Now', 'wc-serial-numbers' ); ?></a>
-			</div>
-			<img src="<?php echo esc_url( WCSN()->assets_url( 'images/add-generator.png' ) ); ?>" alt="<?php esc_attr_e( 'Generators', 'wc-serial-numbers' ); ?>"/>
-		</div>
-		<?php
-	}
-
-	/**
-	 * Debug tab content.
-	 *
-	 * @since 1.4.6
-	 * @return void
-	 */
-	public static function status_tab() {
-		$statuses = array(
-			'Serial Numbers version' => WCSN()->version,
-		);
-		if ( WCSN()->is_pro_active() && function_exists( 'wc_serial_numbers_pro' ) ) {
-			$statuses['Serial Numbers Pro version'] = WCSN_PRO()->get_version();
-		}
-
-		// Check if required tables exist.
-		$required_tables = array(
-			'serial_numbers',
-			'serial_numbers_activations',
-		);
-		foreach ( $required_tables as $table ) {
-			$exists = $GLOBALS['wpdb']->get_var( $GLOBALS['wpdb']->prepare( 'SHOW TABLES LIKE %s', $GLOBALS['wpdb']->prefix . $table ) ); // phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared
-			if ( $exists ) {
-				$statuses[ $table ] = __( 'Table exists', 'wc-serial-numbers' );
-			} else {
-				$statuses[ $table ] = __( 'Table does not exist', 'wc-serial-numbers' );
-			}
-		}
-
-		// Cron jobs.
-		$cron_jobs = array(
-			'wc_serial_numbers_hourly_event' => __( 'Hourly cron', 'wc-serial-numbers' ),
-			'wc_serial_numbers_daily_event'  => __( 'Daily cron', 'wc-serial-numbers' ),
-		);
-		foreach ( $cron_jobs as $cron_job => $cron_job_name ) {
-			$next_scheduled = wp_next_scheduled( $cron_job );
-			if ( $next_scheduled ) {
-				// translators: %s: Next scheduled time.
-				$statuses[ $cron_job_name ] = sprintf( __( 'Next run: %s', 'wc-serial-numbers' ), esc_html( date_i18n( get_option( 'date_format' ) . ' ' . get_option( 'time_format' ), $next_scheduled ) ) );
-			} else {
-				$statuses[ $cron_job_name ] = __( 'Not scheduled', 'wc-serial-numbers' );
-			}
-		}
-		$statuses = apply_filters( 'wc_serial_numbers_plugin_statuses', $statuses );
-		?>
-		<table class="widefat wcsn-status" cellspacing="0" id="wcsn-status">
-			<thead>
-			<tr>
-				<th colspan="3" data-export-label="Serial Numbers"><h2><?php esc_html_e( 'Serial Numbers', 'wc-serial-numbers' ); ?></h2></th>
-			</tr>
-			</thead>
-			<tbody>
-			<?php foreach ( $statuses as $name => $value ) : ?>
-				<tr>
-					<td data-export-label="<?php echo esc_attr( $name ); ?>"><?php echo esc_html( $name ); ?></td>
-					<td class="help">&dash;</td>
-					<td><?php echo esc_html( $value ); ?></td>
-				</tr>
-			<?php endforeach; ?>
-			</tbody>
-
-		</table>
-
-		<?php
-	}
-
-	/**
-	 * Validation section.
-	 *
-	 * @since 1.4.6
-	 * @return void
-	 */
-	public static function api_validation_section() {
-		$args        = array_merge(
-			wcsn_get_products_query_args(),
-			array(
-				'posts_per_page' => - 1,
-				'fields'         => 'ids',
-			)
-		);
-		$the_query   = new \WP_Query( $args );
-		$product_ids = $the_query->get_posts();
-		$products    = array();
-		foreach ( $product_ids as $product_id ) {
-			$product = wc_get_product( $product_id );
-			if ( ! $product ) {
-				continue;
-			}
-			$products[ $product->get_id() ] = sprintf( '%s (#%d)', $product->get_name(), $product->get_id() );
-		}
-
-		Admin::view( 'html-api-validation', array( 'products' => $products ) );
-	}
-
-	/**
-	 * Activation deactivation section.
-	 *
-	 * @since 1.4.6
-	 * @return void
-	 */
-	public static function api_activation_deactivation_section() {
-		$args        = array_merge(
-			wcsn_get_products_query_args(),
-			array(
-				'posts_per_page' => - 1,
-				'fields'         => 'ids',
-			)
-		);
-		$the_query   = new \WP_Query( $args );
-		$product_ids = $the_query->get_posts();
-		$products    = array();
-		foreach ( $product_ids as $product_id ) {
-			$product = wc_get_product( $product_id );
-			if ( ! $product ) {
-				continue;
-			}
-			$products[ $product->get_id() ] = sprintf( '%s (#%d)', $product->get_name(), $product->get_id() );
-		}
-
-		Admin::view( 'html-api-actions', array( 'products' => $products ) );
-	}
-
-	/**
-	 * Stock section.
-	 *
-	 * @since 1.4.6
-	 * @return void
-	 */
-	public static function reports_stock_tab() {
-		Admin::view( 'html-list-stock' );
+		WCSN()->make( Tools::class )->generators_tab();
 	}
 }
