+++ includes/Orders/Orders.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WooCommerceSerialNumbers;
+namespace WooCommerceSerialNumbers\Orders;
 
 use WooCommerceSerialNumbers\B8\Component;
 
@@ -10,9 +10,19 @@ defined( 'ABSPATH' ) || exit;
  * Class Orders.
  *
  * @since   1.0.0
- * @package WooCommerceSerialNumbers
+ * @package WooCommerceSerialNumbers\Orders
  */
 class Orders extends Component {
+
+	/**
+	 * Child components.
+	 *
+	 * @since 2.4.0
+	 * @var array<int|string, class-string>
+	 */
+	public array $components = array(
+		Admin::class,
+	);
 
 	/**
 	 * Register hooks.
