+++ includes/Plugin.php
@@ -20,15 +20,15 @@ class Plugin extends B8\App {
 	 */
 	protected array $components = array(
 		Installer::class,
-		Cron::class,
-		Cache::class,
 		Encryption::class,
 		Cart::class,
-		Orders::class,
-		Stocks::class,
-		Actions::class,
+		Shop::class,
+		Keys\Keys::class,
+		Activations\Activations::class,
+		Orders\Orders::class,
+		Stocks\Stocks::class,
+		Products::class,
 		RestAPI::class,
-		Frontend\Frontend::class,
 		Compat::class,
 		API::class,
 		Admin\Admin::class,
