+++ includes/Products.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WooCommerceSerialNumbers\Admin;
+namespace WooCommerceSerialNumbers;
 
 use WooCommerceSerialNumbers\B8\Component;
 defined( 'ABSPATH' ) || exit; // Exit if accessed directly.
@@ -9,9 +9,19 @@ defined( 'ABSPATH' ) || exit; // Exit if accessed directly.
  * Class Products.
  *
  * @since   1.0.0
- * @package WooCommerceSerialNumbers\Admin
+ * @package WooCommerceSerialNumbers
  */
 class Products extends Component {
+
+	/**
+	 * Whether to load.
+	 *
+	 * @since 2.4.0
+	 * @return bool
+	 */
+	public function autoload(): bool {
+		return is_admin();
+	}
 
 	/**
 	 * Register hooks.
