+++ includes/Stocks/Stocks.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WooCommerceSerialNumbers;
+namespace WooCommerceSerialNumbers\Stocks;
 
 use WooCommerceSerialNumbers\B8\Component;
 use WooCommerceSerialNumbers\Models\Key;
@@ -11,7 +11,7 @@ defined( 'ABSPATH' ) || exit;
  * Class Stocks.
  *
  * @since   1.0.0
- * @package WooCommerceSerialNumbers
+ * @package WooCommerceSerialNumbers\Stocks
  */
 class Stocks extends Component {
 
@@ -28,6 +28,7 @@ class Stocks extends Component {
 		add_action( 'wc_serial_numbers_key_inserted', array( $this, 'update_stocks' ) );
 		add_action( 'wc_serial_numbers_key_updated', array( $this, 'update_stocks' ) );
 		add_action( 'wc_serial_numbers_key_deleted', array( $this, 'update_stocks' ) );
+		add_action( 'wc_serial_numbers_daily_event', array( $this, 'send_stock_alert_email' ) );
 	}
 
 	/**
@@ -81,5 +82,44 @@ class Stocks extends Component {
 		// Update the product stock meta directly.
 		$product->set_stock_quantity( $quantity );
 		$product->save();
+	}
+
+	/**
+	 * Send low stock email notification.
+	 *
+	 * @since 1.2.0
+	 * @return bool
+	 */
+	public function send_stock_alert_email() {
+		if ( 'yes' !== get_option( 'wc_serial_numbers_enable_stock_notification' ) ) {
+			return false;
+		}
+
+		$stock_threshold = get_option( 'wc_serial_numbers_stock_threshold', 5 );
+		$to              = get_option( 'wc_serial_numbers_notification_recipient', get_option( 'admin_email' ) );
+		if ( empty( $to ) ) {
+			return false;
+		}
+
+		$low_stock_products = wcsn_get_stocks_count( $stock_threshold );
+		if ( empty( $low_stock_products ) ) {
+			return false;
+		}
+
+		$subject = __( 'Serial Numbers stock running low', 'wc-serial-numbers' );
+		/** $woocommerce WooCommerce */
+		global $woocommerce;
+		$mailer = $woocommerce->mailer();
+
+		ob_start();
+		wcsn_get_template( 'email-stock-notification.php', array( 'low_stock_products' => $low_stock_products ) );
+		$message = ob_get_contents();
+		ob_get_clean();
+
+		$message = $mailer->wrap_message( $subject, $message );
+		$headers = apply_filters( 'woocommerce_email_headers', '', 'wc_serial_numbers_low_stock_notification', $mailer );
+		$mailer->send( $to, $subject, $message, $headers, array() );
+
+		exit();
 	}
 }
